merge master

+++ lib/cocos2d-x/scripting/lua/cocos2dx_support/Cocos2dxLuaLoader.cpp
@@ -90,7 +90,7 @@ extern "C"
 
             begin = next + 1;
             next = searchpath.find_first_of(";", begin);
-        } while (begin < (int)searchpath.length());
+        }while ( begin < (size_t)searchpath.length());
 
         if (chunk)
         {
